Fix problems when built with 'char' as unsigned
Prefer BYTE and UBYTE for 8-bit ints, rather than using 'char'.
If __STDC__ is defined, then BYTE is defined as 'signed char',
and UBYTE is always defined as 'unsigned char'.

This fixes bugs that arise in the game if signedness of char is
incorrect, including but not limited to the following:
1) you are able to 'take' from enemies multiple times,
2) weapon disappears when character is facing left